+++ web_v2/requests/get_create_poll_requests.go
@@ -5,29 +5,21 @@ import (
 )
 
 const (
-	FilterTypeCreatePollRequestsPermissionType           = "request_details.permission_type"
-	FilterTypeCreatePollRequestsVoteConfirmationRequired = "request_details.vote_confirmation_required"
-	FilterTypeCreatePollRequestsResultProvider           = "request_details.result_provider"
+	FilterTypeCreatePollRequestsPermissionType   = "request_details.permission_type"
+	FilterTypeCreatePollRequestsVoteConfirmation = "request_details.vote_confirmation"
+	FilterTypeCreatePollRequestsResultProvider   = "request_details.result_provider"
 )
 
 var filterTypeCreatePollRequests = map[string]struct{}{
-	FilterTypeCreatePollRequestsPermissionType:           {},
-	FilterTypeCreatePollRequestsVoteConfirmationRequired: {},
-	FilterTypeCreatePollRequestsResultProvider:           {},
-}
-
-const (
-	IncludeTypeCreatePollRequestsResultProvider = "request_details.result_provider"
-)
-
-var includeTypeCreatePollRequests = map[string]struct{}{
-	IncludeTypeCreatePollRequestsResultProvider: {},
+	FilterTypeCreatePollRequestsPermissionType:   {},
+	FilterTypeCreatePollRequestsVoteConfirmation: {},
+	FilterTypeCreatePollRequestsResultProvider:   {},
 }
 
 type GetCreatePollRequestsFilter struct {
 	GetRequestListBaseFilters
 	PermissionType           uint64 `fig:"request_details.permission_type"`
-	VoteConfirmationRequired bool   `fig:"request_details.vote_confirmation_required"`
+	VoteConfirmationRequired bool   `fig:"request_details.vote_confirmation"`
 	ResultProvider           string `fig:"request_details.result_provider"`
 }
 
@@ -41,7 +33,7 @@ func NewGetCreatePollRequests(r *http.Request) (request GetCreatePollRequests, e
 		r,
 		&request.Filters,
 		filterTypeCreatePollRequests,
-		includeTypeCreatePollRequests,
+		map[string]struct{}{},
 	)
 	if err != nil {
 		return request, err